Update usage of nrepl.middleware.print for 0.6.0
In version 0.6.0 of nREPL, the pr-values middleware was replaced with
nrepl.middleware.print.

See https://github.com/nrepl/nREPL/blob/master/CHANGELOG.md
and https://github.com/nrepl/nrepl/pull/118

The new middleware is described as backwards-compatible with
pr-values, but the expected behavior is different.
The new middleware.print expects:

* A namespace-qualified key :nrepl.middleware.print/print to identify
  the printer function

* The printer function has the signature (fn [value writer opts])
  instead of (fn [value opts])

* The printer function writes to a stream instead of returning a
  string

There was an option `:as-html 1` passed to the custom-renderer
function. As far as I can tell this is left over from some older
implementation and is no longer used, because it was always set in the
nREPL middleware stack.