Remove onbeforeunload 'Leave site?' prompt
Setting the onbeforeunload handler causes the browser to pop up a
modal dialog asking:

    Leave site?
    Changes you made may not be saved.

This is useful when Gorilla REPL is used as an editor, but I am using
it with generated source files that I edit elsewhere. I want to ignore
any local changes and skip this prompt when closing a tab/window.